# openapi.yaml
openapi: "3.0.0"
info:
  version: 0.0.1
  title: BBS API
  license:
    name: MIT License
    url: https://choosealicense.com/licenses/mit/
servers:
  - url: http://localhost:8080

paths:
  /boardList:
    get:
      description: Get board list
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardListResponse'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /{server_id}/{board_id}/threadList:
    get:
      description: Get thread list
      parameters:
        - name: server_id
          in: path
          required: true
          schema:
            type: string
        - name: board_id
          in: path
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThreadListResponse'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /{server_id}/{board_id}/thread/{thread_id}:
    get:
      description: Get thread
      parameters:
          - name: server_id
            in: path
            required: true
            schema:
              type: string
          - name: board_id
            in: path
            required: true
            schema:
              type: string
          - name: thread_id
            in: path
            required: true
            schema:
              type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ThreadResponse'
        500:
          description: Internal Server Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'      

components:
  schemas:
    BoardListResponse:
      properties:
        board_groups:
          type: array
          items:
            $ref: "#/components/schemas/BoardGroup"
    BoardGroup:
      properties:
        name:
          type: string
        board_list:
          type: array
          items:
            $ref: "#/components/schemas/Board"
    Board:
      properties:
        name:
          type: string
        server_id:
          type: string
        board_id:
          type: string
    ThreadListResponse:
      properties:
        thread_list:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
    ThreadResponse:
      properties:
        comment_list:
          type: array
          items:
            $ref: "#/components/schemas/Comment"
            type: object
            properties:
              meta:
                type: object
                properties:
                  number:
                    type: integer
                  name:
                    type: string
                  posted_at:
                    type: string
                    format: date-time
                  user_id:
                    type: string
                  comment:
                    type: string
    Comment:
      properties:
        meta:
          type: object
          properties:
            number:
              type: integer
            user_name:
              type: string
            user_id:
              type: string
            posted_at:
              type: string
              format: date-time
        text:
          type: string
    Error:
      required:
        - code
        - message
      properties:
        code:
          type: integer
        message:
          type: string